[#386] Add Blog Posts to the Community Detail Page
This patch adds the CommunityFeed model, along with an associated
InlineAdmin class, allowing staff users to add multiple RSS and Atom
feeds to a community. The latest 5 posts are shown on the Community
Detail Page. The model has a get_blog_posts method which returns the
list of entry dictionaries parsed by feedparser, replacing the
`published` attribute's text with a datetime object.

* Add the get_latest_blog_posts method to the communities.Community
  model to retrieve the 5 latest feed items from all of the Community's
  feeds.
* Add the feedparser dependency to requirements/base.txt.

* Modify core/static/css/fec.css and the communities/details.html
  template to improve the spacing above and below the new `Latest
  Updates` header on the Community Details Page.
* Modify the create_test_image function in the core.utils module so that
  nose does not think it is a test.
* Modify the prospector config file to ignore errors for pyflakes and
  frosted that are better recognized by pylint.

* Cleanup the imports in the communities.models module.