[#509] Add Communities Conference Text
Add a `communities_conference_text` field to the `HomepageContent` model
for customizing the text shown in the `Communities Conference` widget in
the sidebar of general pages.

A SitewideContent model was not created because it would only contain
that one field. One may be added in the future if there is additional
appropriate data.

Refs #509